requirements.txt
absl-py==2.1.0
adk==0.0.5
aenum==3.1.15
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.4
alembic==1.12.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
asttokens==3.0.0
asyncio==3.4.3
attrs==25.1.0
Authlib==1.5.2
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
black==23.7.0
blinker==1.9.0
boto3==1.28.39
botocore==1.31.39
browser-use==0.1.40
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.3.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.1.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
cleo==2.1.0
click==8.1.8
cmake==3.31.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
diffusers==0.32.2
distlib==0.3.9
distro==1.9.0
dlib==19.24.6
dnspython==2.7.0
docstring_parser==0.16
dulwich==0.22.8
durationpy==0.9
easydict==1.13
ecdsa==0.19.1
email_validator==2.2.0
eventlet==0.39.1
executing==2.2.0
face-recognition==1.3.0
# Editable install with no version control (face_recognition_app==0.1)
-e /Users/tomiyasuhiroki/Desktop/CascadeProjects/顔認識2
face_recognition_models==0.3.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
findpython==0.6.3
firebase-admin==6.6.0
firecrawl==1.14.1
flake8==6.1.0
Flask==3.1.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.1.24
fonttools==4.55.6
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-adk==0.1.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.89.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.0
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.30.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.24.0
google-cloud-trace==1.16.1
google-cloud-translate==3.19.0
google-cloud-vision==3.9.0
google-crc32c==1.6.0
google-genai==1.11.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googlemaps==4.5.3
gotrue==1.3.1
graphviz==0.20.3
greenlet==3.1.1
griffe==1.6.0
gritql==0.2.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hatchling==1.27.0
hpack==4.1.0
httpcore==1.0.8
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==7.2.1
importlib_resources==6.5.2
iniconfig==2.0.0
insightface==0.7.3
installer==0.7.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.21
langchain-anthropic==0.3.3
langchain-cli==0.0.36
langchain-community==0.3.20
langchain-core==0.3.54
langchain-google-community==2.0.7
langchain-google-genai==2.1.3
langchain-mcp-adapters==0.0.5
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.7
langdetect==1.0.9
langgraph==0.3.20
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.59
langserve==0.3.1
langsmith==0.3.18
lazy_loader==0.4
line-bot-sdk==3.16.3
loguru==0.7.2
Mako==1.3.9
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mcp==1.6.0
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.1
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.24.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.0
oauthlib==3.2.2
ollama==0.4.7
onnx==1.17.0
onnxruntime==1.20.1
openai==1.75.0
openai-agents==0.0.4
opencv-contrib-python==4.11.0.86
opencv-python==4.8.0.76
opencv-python-headless==4.11.0.86
opentelemetry-api==1.31.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
opt_einsum==3.4.0
orjson==3.10.16
ormsgpack==1.9.0
overrides==7.7.0
paapi5-python-sdk==1.1.0
packaging==24.2
pandas==2.1.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pbs-installer==2025.4.9
pdf2image==1.17.0
# Editable install with no version control (persona_interview_agent==0.1.0)
-e /Users/tomiyasuhiroki/Desktop/persona_interview_project
pexpect==4.9.0
pillow==10.4.0
pillow_heif==0.22.0
pkginfo==1.12.1.2
platformdirs==4.3.6
playwright==1.51.0
plotly==5.18.0
pluggy==1.5.0
poetry==2.1.2
poetry-core==2.1.2
postgrest==0.10.8
posthog==3.21.0
prettytable==3.14.0
proglog==0.1.10
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.25.0
protobuf==5.29.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.0
pydantic_core==2.33.1
pydeck==0.9.1
pyee==12.1.1
pyflakes==3.1.0
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.10
pytest==7.4.0
pytest-flask==1.2.0
pytest-mock==3.14.0
python-amazon-paapi==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.11.2
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.12.1
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.13.0
realtime==1.0.6
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-aws4auth==1.3.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.6.2
safetensors==0.5.2
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.8.0
shapely==2.1.0
shellingham==1.5.4
simple-websocket==1.1.0
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.23
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.46.2
storage3==0.5.4
streamlit==1.43.2
StrEnum==0.4.15
stringzilla==3.11.3
supabase==1.0.4
supafunc==0.2.3
sympy==1.13.3
tenacity==8.2.3
threadpoolctl==3.5.0
tifffile==2025.1.10
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
transformers==4.48.1
trove-classifiers==2025.4.11.15
typer==0.15.2
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.1
uvloop==0.21.0
validators==0.34.0
virtualenv==20.30.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xattr==1.1.4
xlrd==2.0.1
xxhash==3.5.0
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
